JSファイルとHTMLファイルの分割と内容の更新
