+++ .github/workflows/main.yml
@@ -21,6 +21,8 @@ jobs:
         node-version: 16.x
     - name: Install NPM packages
       run: npm ci
+    - name: Compile HTML
+      run: npm run compile
     - name: Build website
       run: npm run build --if-present
     - name: Deploy website
