+++ docs/package.json
@@ -7,10 +7,10 @@
     "dev": "webpack --mode=development",
     "build": "webpack --mode=production",
     "watch": "webpack --watch",
-    "start": "webpack serve --open --mode=development",
-		"compile": "html-includes --src src/html --dest dist",
-		"compile:watch": "html-includes --src src/html --dest dist --watch",
-		"compile:min": "html-includes --src src --dest dist --minify"
+    "start": "webpack serve --open --mode=development --analyze",
+    "compile": "html-includes --src src/html --dest dist",
+    "compile:watch": "html-includes --src src/html --dest dist --watch",
+    "compile:min": "html-includes --src src --dest dist --minify"
   },
   "keywords": [],
   "author": "",
@@ -27,14 +27,14 @@
     "html-webpack-plugin": "^5.5.0",
     "style-loader": "^3.2.1",
     "webpack": "^5.53.0",
+    "webpack-bundle-analyzer": "^4.5.0",
     "webpack-cli": "^4.8.0",
     "webpack-dev-server": "^4.2.1"
   },
   "dependencies": {
     "bindings": "^1.5.0",
+    "clipboard": "^2.0.11",
     "dark-mode-switch": "^1.0.0",
-    "formdata-polyfill": "^4.0.7",
-    "nan": "^2.15.0",
-    "whatwg-fetch": "^3.6.2"
+    "nan": "^2.15.0"
   }
 }
