+++ docs/src/html/BLEProblem.html
@@ -1,19 +1,22 @@
 <!DOCTYPE html>
 <html lang="ja">
-<head>
-	${require('/common/_head.html') pageTitle="BLE Micro Pro に関する問題"}
-</head>
-<body class="bg-light">
-  <div class="container" id="container">
-		${require('/common/_header.html')}
+	<head>
+		${require('/common/_head.html') pageTitle="BLE Micro Pro に関する問題"}
+	</head>
+	<body class="bg-light">
+	  <div class="container" id="container">
+			${require('/common/_header.html')}
 
-    <div class="border rounded m-3 p-3 bg-white shadow-sm">
-      <h2 class="mb-3">BLE Micro Pro に関する問題の場合</h2>
-      <p class="lh-base">BLE Micro Pro の問題については、Discord の Self-Made Keyboards in Japan の #ble-micro-pro で質問してください。<br>
-      質問するときは、現在発生している問題や PC 環境などを具体的に説明した上で質問してください。</p>
-      <a class="btn btn-primary" href="https://discord.gg/zXCss8T" role="button">Discord の Self-Made Keyboards in Japan に参加する</a>
-    </div>
+	    <div class="border rounded m-3 p-3 bg-white shadow-sm">
+	      <h2 class="mb-3">BLE Micro Pro に関する問題の場合</h2>
+	      <p class="lh-base">BLE Micro Pro の問題については、Discord の Self-Made Keyboards in Japan の #ble-micro-pro で質問してください。<br>
+	      質問するときは、現在発生している問題や PC 環境などを具体的に説明した上で質問してください。</p>
+				<a class="btn btn-primary fs-6 px-2 text-start" href="https://discord.gg/zXCss8T" role="button">
+					<i class="fab fa-discord me-2"></i>
+					Discord の Self-Made Keyboards in Japan に参加する</a>
+	    </div>
 
-  <link rel="stylesheet" href="css/style.css">
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
+		${require('/common/_footer.html')}
+		</div>
+	</body>
 </html>
