+++ docs/src/html/common/_firmwareInfo.html
@@ -1,43 +1,58 @@
 		<h2>ファームウェアの情報</h2>
-			<div class="form-group p-2 bg-white" id="firmwareInfomation">
-        <label>ファームウェアの種類</label>
-        <div class="form-check">
-          <input class="form-check-input" type="radio" name="firmwareName" id="QMKFirmware" value="QMK Firmware">
-          <label class="form-check-label" for="QMKFirmware">QMK Toolbox</label>
-        </div>
-        <div class="form-check">
-          <input class="form-check-input" type="radio" name="firmwareName" id="VIA" value="VIA (Remap)">
-          <label class="form-check-label" for="VIA">VIA (Remap)</label>
-        </div>
-        <div class="form-check">
-          <input class="form-check-input" type="radio" name="firmwareName" id="KMKFirmware" value="KMK Firmware">
-          <label class="form-check-label" for="KMKFirmware">KMK Firmware</label>
-        </div>
-        <div class="form-check">
-          <input class="form-check-input" type="radio" name="firmwareName" id="PRKFirmware" value="PRK Firmware">
-          <label class="form-check-label" for="PRKFirmware">PRK Firmware</label>
-        </div>
-        <div class="form-check">
-          <input class="form-check-input" type="radio" name="firmwareName" id="otherFirmware" value="otherFirmware">
-          <label class="form-check-label" for="otherFirmware">その他のファームウェア</label>
-        </div>
-				<div class="form-group p-2 bg-white">
-					<label>ファームウェア書き込みツール</label>
-					<div class="form-check">
-						<input class="form-check-input" type="radio" name="writingTool" id="qmkToolbox" value="QMK Toolbox">
-						<label class="form-check-label" for="qmkToolbox">QMK Toolbox</label>
-					</div>
-					<div class="form-check">
-						<input class="form-check-input" type="radio" name="writingTool" id="qmkCommandline"
-							value="make コマンドでビルド＆書き込み">
-						<label class="form-check-label" for="qmkCommandline"><code>make</code> or <code>qmk flash</code>
-							コマンドでビルド＆書き込み</label>
-					</div>
-					<div class="form-check">
-						<input class="form-check-input" type="radio" name="writingTool" id="proMicroWebUpdater" value="proMicroWebUpdater">
-						<label class="form-check-label" for="proMicroWebUpdater">Pro Micro Web Updater</label>
-						<a href="https://sekigon-gonnoc.github.io/promicro-web-updater/index.html">( link )</a>
-					</div>
+		<div class="form-group p-2 bg-white" id="firmwareInfomation">
+      <label>ファームウェアの種類</label>
+      <div class="form-check">
+        <input class="form-check-input" type="radio" name="firmware" id="QMKFirmware" value="QMK Firmware">
+        <label class="form-check-label" for="QMKFirmware">QMK Toolbox</label>
+      </div>
+      <div class="form-check">
+        <input class="form-check-input" type="radio" name="firmware" id="VIA" value="VIA (Remap)">
+        <label class="form-check-label" for="VIA">VIA (Remap)</label>
+      </div>
+      <div class="form-check">
+        <input class="form-check-input" type="radio" name="firmware" id="KMKFirmware" value="KMK Firmware">
+        <label class="form-check-label" for="KMKFirmware">KMK Firmware</label>
+      </div>
+      <div class="form-check">
+        <input class="form-check-input" type="radio" name="firmware" id="PRKFirmware" value="PRK Firmware">
+        <label class="form-check-label" for="PRKFirmware">PRK Firmware</label>
+      </div>
+      <div class="form-check">
+        <input class="form-check-input" type="radio" name="firmware" id="otherFirmware" value="その他のファームウェア">
+        <label class="form-check-label" for="otherFirmware">その他のファームウェア</label>
+      </div>
+			<div class="form-group p-2">
+				<div class="form-outline">
+					<input class="form-control" type="text" name="firmwareName" id="otherFirmwareInput" disabled>
+					<label class="form-label" for="otherFirmwareInput">ファームウェア名</label>
 				</div>
 			</div>
-  
+		</div>
+		<div class="form-group p-2 bg-white" id="writingToolInformation">
+			<label>ファームウェア書き込みツール</label>
+			<div class="form-check">
+				<input class="form-check-input" type="radio" name="writingTool" id="qmkToolbox" value="QMK Toolbox">
+				<label class="form-check-label" for="qmkToolbox">QMK Toolbox</label>
+			</div>
+			<div class="form-check">
+				<input class="form-check-input" type="radio" name="writingTool" id="qmkCommandline"
+					value="make コマンドでビルド＆書き込み">
+				<label class="form-check-label" for="qmkCommandline"><code>make</code> or <code>qmk flash</code>
+					コマンドでビルド＆書き込み</label>
+			</div>
+			<div class="form-check">
+				<input class="form-check-input" type="radio" name="writingTool" id="proMicroWebUpdater" value="proMicroWebUpdater">
+				<label class="form-check-label" for="proMicroWebUpdater">Pro Micro Web Updater</label>
+				<a href="https://sekigon-gonnoc.github.io/promicro-web-updater/index.html">( link )</a>
+			</div>
+			<div class="form-check">
+				<input class="form-check-input" type="radio" name="writingTool" id="otherMethod" value="その他のツール">
+				<label class="form-check-label" for="otherMethod">その他の書き込み方法</label>
+			</div>
+			<div class="form-group p-2">
+				<div class="form-outline">
+					<input class="form-control" type="text" name="writingToolName" id="otherMethodInput" disabled>
+					<label class="form-label" for="otherMethodInput">書き込み方法</label>
+				</div>
+			</div>
+		</div>
