+++ docs/src/html/common/_footer.html
@@ -1,0 +1,6 @@
+<link rel="stylesheet" href="css/style.css">
+<!-- MDB -->
+<script
+type="text/javascript"
+src="https://cdnjs.cloudflare.com/ajax/libs/mdb-ui-kit/4.0.0/mdb.min.js"
+></script>
