+++ docs/src/html/common/_header.html
@@ -1,21 +1,58 @@
-		<div class="row m-3">
-			<div class="col p-0">
-				<div class="custom-control custom-switch">
-					<input type="checkbox" class="custom-control-input" id="darkSwitch" />
-					<label class="custom-control-label" for="darkSwitch">Dark Mode</label>
-				</div>
-	    </div>
-			<div class="col p-0 text-end">
-		    <p class="text-right m-0">
-		      <a href="index_en.html" class="btn btn-outline-info">English Version</a>
-	  	  </p>
+<div class="row m-3">
+	<div class="col p-0 text-end">
+		<p class="text-right m-0">
+			<a href="index_en.html" class="btn btn-outline-info">English Version</a>
+		</p>
+	</div>
+</div>
+
+<header>
+	<!--	 Navbar -->
+	<nav class="navbar navbar-expand-lg navbar-light bg-white mx-3">
+		<div class="container">
+			<button
+			class="navbar-toggler"
+			type="button"
+			data-mdb-toggle="collapse"
+			data-mdb-target="#navbar"
+			aria-controls="navbar"
+			aria-expanded="false"
+			aria-label="Toggle navigation"
+			>
+				<i class="fas fa-bars"></i>
+			</button>
+			<div class="collapse navbar-collapse" id="navbar">
+				<ul class="navbar-nav me-auto mb-1 mb-lg-0">
+					<li class="nav-item border-end">
+						<a class="nav-link text-dark" aria-current="page" href="index.html">Home</a>
+					</li>
+					<li class="nav-item border-end">
+						<a class="nav-link text-dark" href="buildProblem.html">組み立ての問題</a>
+					</li>
+					<li class="nav-item border-end">
+						<a class="nav-link text-dark" href="firmwareProblem.html">ファームウェアの問題</a>
+					</li>
+					<li class="nav-item border-end">
+						<a class="nav-link text-dark" href="BLEProblem.html">BLE Micro Proの問題</a>
+					</li>
+					<li class="nav-item border-end">
+						<a class="nav-link text-dark" href="designProblem.html">設計の問題</a>
+					</li>
+					<li class="nav-item border-end">
+						<a class="nav-link text-dark" href="otherProblem.html">その他の問題</a>
+					</li>
+				</ul>
 			</div>
-	  </div>
-		<div class="border rounded m-3 p-3 bg-primary shadow-sm" style="--bs-bg-opacity: .5;">
-		  <h1 class="text-center mb-5 text-white">問診票テンプレート</h1>
-		  <p class="lead text-white">
-		    Discord の #mon-shin チャンネルに投稿する際の問診票です。<br>
-		    質問に沿って回答していくと投稿用の文章が作成されます。<br>
-		    動作に不具合があったときは、Discord の #keyboard-zatsu チャンネルで知らせてください。
-		  </p>
 		</div>
+	</nav>
+	<!-- Navbar -->
+
+	<div class="border-start border-end border-bottom mb-3 mx-3 p-3 bg-white">
+	  <h1 class="text-center mb-5">問診票テンプレート</h1>
+	  <p class="lead">
+	    Discord の #mon-shin チャンネルに投稿する際の問診票です。<br>
+	    質問に沿って回答していくと投稿用の文章が作成されます。<br>
+	    動作に不具合があったときは、Discord の #keyboard-zatsu チャンネルで知らせてください。
+	  </p>
+	</div>
+</header>
