+++ docs/src/html/common/_otherInfo.html
@@ -1,13 +1,17 @@
         <h2 class="">その他の情報</h2>
         <div class="form-group p-2" id="detailWork">
-          <label for="detailWork">これまでに行った作業の内容</label>
-          <textarea class="form-control" rows="3" name="detailWork" id="detailWorkTextarea"></textarea>
+					<div class="form-outline">
+	          <textarea class="form-control" rows="3" name="detailWork" id="detailWorkTextarea"></textarea>
+	          <label class="form-label" for="detailWork">これまでに行った作業の内容</label>
+					</div>
         </div>
         <div class="form-group p-2">
-          <label for="informationReferredTo">参考にした情報（ビルドガイド除く）</label>
-          <textarea class="form-control" type="text" placeholder="https://" rows="3" name="informationReferredTo"
-            id="informationReferredTo"></textarea>
-        </div>
+					<div class="form-outline">
+	          <textarea class="form-control" type="text" placeholder="https://" rows="3" name="informationReferredTo"
+	            id="informationReferredTo"></textarea>
+						<label class="form-label" for="informationReferredTo">参考にした情報（ビルドガイド除く）</label>
+					</div>
+					</div>
         <div class="form-group p-2" id="haveTester">
           <label>テスターを持っていますか？</label>
           <div class="form-check">
