+++ docs/src/html/common/_testMicrocomputerOnly.html
@@ -2,11 +2,11 @@
           <label for="microcomputeronly">マイコンボードだけ PC に接続したらキーボードとして認識されますか？</label>
           <div class="form-check">
             <input class="form-check-input" type="radio" name="testMicrocomputerOnly" id="microcomputerOnlyYes" value="はい">
-            <label class="form-check-label" for="promicroOnlyYes">はい</label>
+            <label class="form-check-label" for="microcomputerOnlyYes">はい</label>
           </div>
           <div class="form-check">
             <input class="form-check-input" type="radio" name="testMicrocomputerOnly" id="microcomputerOnlyNo" value="いいえ">
-            <label class="form-check-label" for="promicroOnlyNo">いいえ</label>
+            <label class="form-check-label" for="microcomputerOnlyNo">いいえ</label>
           </div>
           <div class="form-check">
             <input class="form-check-input" type="radio" name="testMicrocomputerOnly" id="noUseConthrough" value="コンスルー未使用">
