+++ docs/src/html/designProblem.html
@@ -1,38 +1,58 @@
 <!DOCTYPE html>
 <html lang="ja">
-<head>
-	${require('/common/_head.html') pageTitle="設計に関する問題"}
-</head>
-<body class="bg-light">
-  <div class="container" id="container">
-		${require('/common/_header.html')}
+	<head>
+		${require('/common/_head.html') pageTitle="設計に関する問題"}
+	</head>
+	<body class="bg-light">
+	  <div class="container" id="container">
+			${require('/common/_header.html')}
 
-    <div class="border rounded m-3 p-3 bg-white shadow-sm">
-      <div class="card-group">
-        <div class="card">
-          <div class="card-body">
-            <h4 class="card-title text-decoration-underline">設計初心者の方々</h4>
-            <p class="card-text">キーボードの設計が初めてで分からないことがたくさんある方などは、Discord の Self-Made Keyboards in Japan の #development-beginner チャンネルへ</p>
-            <a class="btn btn-primary" href="https://discord.gg/zXCss8T" role="button">Discord の Self-Made Keyboards in Japan に参加する</a>
-          </div>
-        </div>
-        <div class="card">
-          <div class="card-body">
-            <h4 class="card-title text-decoration-underline">実装に関する問題や相談したいことがある方々</h4>
-            <p class="card-text">部品の選定基準や回路設計の相談のほか、新しい実装方法の提案などは、Discord の Self-Made Keyboards in Japan の #development-lobby チャンネルへ</p>
-            <a class="btn btn-primary" href="https://discord.gg/zXCss8T" role="button">Discord の Self-Made Keyboards in Japan に参加する</a>
-          </div>
-        </div>
-        <div class="card">
-          <div class="card-body">
-            <h4 class="card-title text-decoration-underline">その他の問題や相談</h4>
-            <p class="card-text">上記以外の問題や、上記のどちらに当てはまるか分からない場合は、Discord の Self-Made Keyboards in Japan の #keybord-zatsu チャンネルへ</p>
-            <a class="btn btn-primary" href="https://discord.gg/zXCss8T" role="button">Discord の Self-Made Keyboards in Japan に参加する</a>
-          </div>
-        </div>
-      </div>
-    </div>
+	    <div class="border rounded m-3 p-3 bg-white shadow-sm row row-cols-md-3">
+	      <div class="col">
+	        <div class="card border h-100">
+	          <div class="card-body">
+	            <h4 class="card-title text-decoration-underline">設計初心者の方々</h4>
+	            <p class="card-text">キーボードの設計が初めてで分からないことがたくさんある方などは、Discord の Self-Made Keyboards in Japan の #development-beginner チャンネルへ</p>
+	          </div>
+						<div class="card-footer">
+	            <a class="btn btn-primary fs-6 px-2 text-start" href="https://discord.gg/zXCss8T" role="button">
+								<i class="fab fa-discord me-2"></i>
+								Discord の Self-Made Keyboards in Japan に参加する
+							</a>
+						</div>
+	        </div>
+				</div>
+	      <div class="col">
+	        <div class="card border h-100">
+	          <div class="card-body">
+	            <h4 class="card-title text-decoration-underline">実装に関する問題や相談したいことがある方々</h4>
+	            <p class="card-text">部品の選定基準や回路設計の相談のほか、新しい実装方法の提案などは、Discord の Self-Made Keyboards in Japan の #development-lobby チャンネルへ</p>
+						</div>
+						<div class="card-footer">
+	            <a class="btn btn-primary fs-6 px-2 text-start" href="https://discord.gg/zXCss8T" role="button">
+								<i class="fab fa-discord me-2"></i>
+								Discord の Self-Made Keyboards in Japan に参加する
+							</a>
+						</div>
+	        </div>
+				</div>
+	      <div class="col">
+	        <div class="card border h-100">
+	          <div class="card-body">
+	            <h4 class="card-title text-decoration-underline">その他の問題や相談</h4>
+	            <p class="card-text">上記以外の問題や、上記のどちらに当てはまるか分からない場合は、Discord の Self-Made Keyboards in Japan の #keybord-zatsu チャンネルへ</p>
+	          </div>
+						<div class="card-footer">
+	            <a class="btn btn-primary fs-6 px-2 text-start" href="https://discord.gg/zXCss8T" role="button">
+								<i class="fab fa-discord me-2"></i>
+								Discord の Self-Made Keyboards in Japan に参加する
+							</a>
+						</div>
+	        </div>
+	      </div>
+	    </div>
 
-  <link rel="stylesheet" href="css/style.css">
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
+			${require('/common/_footer.html')}
+		</div>
+	</body>
 </html>
