+++ docs/src/html/index.html
@@ -8,39 +8,38 @@
 		${require('/common/_header.html')}
 
     <div class="m-0 p-3">
-      <div class="border rounded mt-1 mb-1 p-3 bg-white shadow-sm">
-        <p class="fw-bold">組み立てに関する問題</p>
-        <a class="btn btn-outline-primary" href="./buildProblem.html" role="button">
-          組み立てたけど入力が出来ない、LEDが点灯しない 等の問題が起きている場合はこちらのページへ
+      <div class="border rounded mt-1 mb-1 p-3 bg-light shadow-sm">
+        <p class="fw-bold fs-5">組み立てに関する問題</p>
+        <a class="btn btn-primary text-white fs-6" href="./buildProblem.html" role="button">
+          組み立てたけど入力が出来ない、LEDが点灯しないなどの問題はこちらのページへ
         </a>
       </div>
-      <div class="border rounded mt-1 mb-1 p-3 bg-white shadow-sm">
-        <p class="fw-bold">ファームウェアに関する問題</p>
-        <a class="btn btn-outline-primary" href="./firmwareProblem.html" role="button">
+      <div class="border rounded mt-1 mb-1 p-3 bg-light shadow-sm">
+        <p class="fw-bold fs-5">ファームウェアに関する問題</p>
+        <a class="btn btn-primary text-white fs-6" href="./firmwareProblem.html" role="button">
           ファームウェアの書き込み、ビルド、環境構築の問題はこちらのページへ
         </a>
       </div>
-      <div class="border rounded mt-1 mb-1 p-3 bg-white shadow-sm">
-        <p class="fw-bold">BLE Micro Pro に関する問題</p>
-        <a class="btn btn-outline-primary" href="./BLEProblem.html" role="button">
-          BLE Micro Pro を使っている時に Bluetooth 接続が出来ないなどの問題はこちらのページへ
+      <div class="border rounded mt-1 mb-1 p-3 bg-light shadow-sm">
+        <p class="fw-bold fs-5">BLE Micro Pro に関する問題</p>
+        <a class="btn btn-primary text-white fs-6" href="./BLEProblem.html" role="button">
+          BLE Micro Pro で Bluetooth 接続が出来ないなどの問題はこちらのページへ
         </a>
       </div>
-      <div class="border rounded mt-1 mb-1 p-3 bg-white shadow-sm">
-        <p class="fw-bold">設計に関する問題</p>
-          <a class="btn btn-outline-primary" href="./designProblem.html" role="button">
-            キーボード設計に関する相談や質問についてはこちらのページへ
+      <div class="border rounded mt-1 mb-1 p-3 bg-light shadow-sm">
+        <p class="fw-bold fs-5">設計に関する問題</p>
+          <a class="btn btn-primary text-white fs-6" href="./designProblem.html" role="button">
+            キーボード設計に関する相談や質問はこちらのページへ
           </a>
       </div>
-      <div class="border rounded mt-1 mb-1 p-3 bg-white shadow-sm">
-        <p class="fw-bold">その他の問題</p>
-        <a class="btn btn-outline-primary" href="./otherProblem.html" role="button">
+      <div class="border rounded mt-1 mb-1 p-3 bg-light shadow-sm">
+        <p class="fw-bold fs-5">その他の問題</p>
+        <a class="btn btn-primary text-white fs-6" href="./otherProblem.html" role="button">
           上記以外の問題はこちらのページへ
         </a>
       </div>
     </div>
   </div>
 
-  <link rel="stylesheet" href="css/style.css">
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
+	${require('/common/_footer.html')}
 </html>
