+++ docs/src/html/otherProblem.html
@@ -1,19 +1,23 @@
 <!DOCTYPE html>
 <html lang="ja">
-<head>
-	${require('/common/_head.html') pageTitle="その他の問題に関する問題"}
-</head>
-<body class="bg-light">
-  <div class="container" id="container">
-		${require('/common/_header.html')}
+	<head>
+		${require('/common/_head.html') pageTitle="その他の問題に関する問題"}
+	</head>
+	<body class="bg-light">
+	  <div class="container" id="container">
+			${require('/common/_header.html')}
 
-    <div class="border rounded m-3 p-3 bg-white shadow-sm">
-      <h2 class="mb-3">その他の問題の場合</h2>
-      <p class="lh-base">設計・組み立て・ファームウェア・BLE Micro Pro 以外の問題については、ひとまず Discord の Self-Made Keyboards in Japan の #keyboard-zatsu で質問してください。<br>
-      質問するときは、抱えている問題の詳細などを説明してください。</p>
-      <a class="btn btn-primary" href="https://discord.gg/zXCss8T" role="button">Discord の Self-Made Keyboards in Japan に参加する</a>
-    </div>
+	    <div class="border rounded m-3 p-3 bg-white shadow-sm">
+	      <h2 class="mb-3">その他の問題の場合</h2>
+	      <p class="lh-base">設計・組み立て・ファームウェア・BLE Micro Pro 以外の問題については、ひとまず Discord の Self-Made Keyboards in Japan の #keyboard-zatsu で質問してください。<br>
+	      質問するときは、抱えている問題の詳細などを説明してください。</p>
+				<a class="btn btn-primary fs-6 px-2" href="https://discord.gg/zXCss8T" role="button">
+					<i class="fab fa-discord me-2"></i>
+					Discord の Self-Made Keyboards in Japan に参加する
+				</a>
+			</div>
 
-  <link rel="stylesheet" href="css/style.css">
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
+		${require('/common/_footer.html')}
+		</div>
+	</body>
 </html>
